requirements.txt
aenum-2.2.3
dataclasses==0.7
numpy==1.16.2
